ubiaiback/requirements.txt
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
